+++ app/src/main/java/com/nprotech/moneytracker/ui/activities/CurrencyPickerActivity.java
@@ -104,7 +104,7 @@ public class CurrencyPickerActivity extends BaseActivity {
             if (type.equalsIgnoreCase("account") || type.equalsIgnoreCase("goal")) {
                 currencyLists = masterViewModel.getAllCurrencies();
             } else if (type.equalsIgnoreCase("wallet")) {
-                currencyLists = masterViewModel.getCurrenciesForWallet((int) PreferenceManager.INSTANCE.getAccountId());
+                currencyLists = masterViewModel.getCurrenciesForWallet(PreferenceManager.INSTANCE.getAccountId());
             }
 
             if (!currencyLists.isEmpty()) {
