+++ app/src/main/java/com/nprotech/moneytracker/ui/fragments/InProgressGoalFragment.java
@@ -87,7 +87,7 @@ public class InProgressGoalFragment extends Fragment {
     private void bindData() {
         try {
 
-            goalViewModel.getGoals((int) PreferenceManager.INSTANCE.getAccountId(), false, false).observe(getViewLifecycleOwner(), goalWithDetails -> {
+            goalViewModel.getGoals(PreferenceManager.INSTANCE.getAccountId(), false, false).observe(getViewLifecycleOwner(), goalWithDetails -> {
                 if (goalWithDetails.isEmpty()) {
                     emptyWrapper.setVisibility(View.VISIBLE);
                     rvGoals.setVisibility(View.GONE);
